requirements.txt
asgiref==3.3.1
bootstrap4==0.1.0
certifi==2020.11.8
chardet==3.0.4
DAWG-Python==0.7.2
Django==3.1.3
django-ckeditor==6.0.0
django-crispy-forms==1.9.2
django-js-asset==1.2.2
django-tinymce==3.1.0
docopt==0.6.2
docxtpl==0.11.2
gitdb==4.0.5
GitPython==3.1.11
idna==2.10
Jinja2==2.11.2
lxml==4.6.1
MarkupSafe==1.1.1
numpy==1.19.4
pandas==1.1.4
pep8==1.7.1
Pillow==8.0.1
pymorphy2==0.9.1
pymorphy2-dicts-ru==2.4.417127.4579844
python-dateutil==2.8.1
python-docx==0.8.10
pytz==2020.4
requests==2.25.0
six==1.15.0
smmap==3.0.4
sqlparse==0.4.1
urllib3==1.26.2
djangorestframework==3.12.4
markdown==3.3.4
django-filter==2.4.0
djoser==2.1.0